% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_val}
\alias{get_val}
\title{Get value from environmental variable or R option or default in that order if the previous one is missing.}
\usage{
get_val(env_var_name, option_name, default = NULL)
}
\arguments{
\item{env_var_name}{(\code{character(1)}) name of the system variable}

\item{option_name}{(\code{character(1)}) name of the option}

\item{default}{optional, default value if both \code{Sys.getenv(env_var_name)} and \code{getOption(option_name)} are empty}
}
\value{
an object of any class
keywords internal
}
\description{
Get value from environmental variable or R option or default in that order if the previous one is missing.
}
