% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_handlers.R
\name{register_handlers_possible}
\alias{register_handlers_possible}
\title{Checks if it is possible to register handlers.}
\usage{
register_handlers_possible()
}
\value{
boolean
}
\description{
Checks if it is possible to register handlers.
}
\examples{
register_handlers_possible() # TRUE
try(register_handlers_possible()) # FALSE
withCallingHandlers(register_handlers_possible()) # FALSE

\dontrun{
# To avoid errors when registering:
try(globalCallingHandlers(NULL))
withCallingHandlers(globalCallingHandlers(NULL), foo = identity)
}
}
\keyword{internal}
