% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supress_logs.R
\name{suppress_logs}
\alias{suppress_logs}
\title{Suppress logger logs}
\usage{
suppress_logs()
}
\value{
\code{NULL} invisible
}
\description{
This function suppresses \code{logger} when running tests via \code{testthat}.
To suppress logs for a single test, add this function
call within the \code{testthat::test_that} expression. To suppress logs for an entire
test file, call this function at the start of the file.
}
\examples{
testthat::test_that("An example test", {
  suppress_logs()
  testthat::expect_true(TRUE)
})

}
