% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_formatter.R
\name{teal_logger_formatter}
\alias{teal_logger_formatter}
\alias{teal_logger_transformer}
\title{Teal \code{log_formatter}}
\usage{
teal_logger_formatter()

teal_logger_transformer(text, envir)
}
\arguments{
\item{text}{Text (typically) R code to parse and evaluate.}

\item{envir}{environment to evaluate the code in}
}
\value{
Nothing. Called for its side effects.
}
\description{
Custom \code{log_formatter} supporting atomic vectors. By default \code{glue::glue}
returns n-element vector when vector is provided as an input. This function
generates \code{"[elem1, elem2, ...]"} for atomic vectors. Function also handles
\code{NULL} value which normally causes \code{logger} to return empty character.
}
\keyword{internal}
