% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_shiny_input_changes.R
\name{log_shiny_input_changes}
\alias{log_shiny_input_changes}
\title{Auto logging input changes in Shiny app}
\usage{
log_shiny_input_changes(
  input,
  level = logger::INFO,
  namespace = NA_character_,
  excluded_inputs = character(),
  excluded_patterns = character(),
  skip_init = FALSE
)
}
\arguments{
\item{input}{passed from Shiny's \code{server}}

\item{level}{log level}

\item{namespace}{the name of the namespace}

\item{excluded_inputs}{character vector of input names to exclude from logging}

\item{excluded_patterns}{character of length one including a grep pattern of names to be excluded from logging}

\item{skip_init}{should initialization message be displayed}
}
\description{
This is to be called in the \code{server} section of the Shiny app.
}
\details{
Function taken from \pkg{logger} package and improved in this PR
\href{https://github.com/daroczig/logger/pull/155/}{daroczig/logger/pull/155}. Once this PR gets merged, the function will
be removed.
}
\examples{
\dontrun{
library(shiny)

ui <- bootstrapPage(
  numericInput("mean1", "mean1", 0),
  numericInput("mean2", "mean2", 0),
  numericInput("sd", "sd", 1),
  textInput("title", "title", "title"),
  textInput("foo", "This is not used at all, still gets logged", "foo"),
  passwordInput("password", "Password not to be logged", "secret"),
  plotOutput("plot")
)

server <- function(input, output) {
  log_shiny_input_changes(input, excluded_inputs = "password", exclude_patterns = "mean")

  output$plot <- renderPlot({
    hist(rnorm(1e3, input$mean, input$sd), main = input$title)
  })
}

shinyApp(ui = ui, server = server)
}
}
\keyword{internal}
