% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{register_handlers}
\alias{register_handlers}
\title{Register handlers for logging messages, warnings and errors}
\usage{
register_handlers(namespace, package = namespace)
}
\arguments{
\item{namespace}{(\code{character(1)}) the logger namespace}

\item{package}{(\code{character(1)}) the package name}
}
\value{
\code{invisible(NULL)}
}
\description{
Register handlers for logging messages, warnings and errors
}
\details{
This function registers global handlers for messages, warnings and errors.
The handlers will investigate the call stack and if it contains a function
from the package, the message, warning or error will be logged into the respective
namespace.
}
\examples{
\dontrun{
register_logger(namespace = "teal.logger")
register_handlers("teal.logger")
}
}
\seealso{
\code{\link[=globalCallingHandlers]{globalCallingHandlers()}}
}
