% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supress_logs.R
\name{suppress_logs}
\alias{suppress_logs}
\title{Suppress logger logs}
\usage{
suppress_logs()
}
\value{
\code{NULL} invisible
}
\description{
This function will suppress the logger when running tests
in order to suppress logs for a single test add this function
call within the test_that function. To suppress logs for an entire
test file, call this function at the start of the file
}
\keyword{internal}
