% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.logger-package.R
\docType{package}
\name{teal.logger-package}
\alias{teal.logger}
\alias{teal.logger-package}
\title{teal.logger: Logging Setup for the 'teal' Family of Packages}
\description{
Logging Setup for the \code{teal} Family of Packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/teal.logger/}
  \item \url{https://github.com/pharmaverse/teal.logger/}
  \item Report bugs at \url{https://github.com/pharmaverse/teal.logger/issues}
}

}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Konrad Pagacz
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
