% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.logger-package.R
\docType{package}
\name{teal.logger-package}
\alias{teal.logger}
\alias{teal.logger-package}
\title{teal.logger: Logging Setup For The `teal` Family Of Packages}
\description{
Logging setup for the \code{teal} family of packages.
}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

Authors:
\itemize{
  \item Nikolas Burkoff \email{nikolas.burkoff@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
  \item Pawel Rucki \email{pawel.rucki@contractors.roche.com}
  \item Junlue Zhao \email{zhaoj88@gene.com}
}

Other contributors:
\itemize{
  \item Tadeusz Lewandowski \email{tadeusz.lewandowski@roche.com} [contributor]
  \item Roche [copyright holder, funder]
}

}
\keyword{internal}
