% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.logger-package.R
\docType{package}
\name{teal.logger-package}
\alias{teal.logger}
\alias{teal.logger-package}
\title{teal.logger: Logging Setup For The `teal` Family Of Packages}
\description{
Logging setup for the \code{teal} family of packages.
}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
